requirements.txt
streamlit
face_recognition
opencv-python-headless
pandas
Pillow
numpy
